backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
pinecone-client
unstructured
pdf2image
pytesseract
pypdf
celery
redis
python-docx
reportlab
python-jose[cryptography]
passlib[bcrypt]
slowapi
sentry-sdk[fastapi]
pytest
httpx
